voice-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
